feat: show target price based on desired yield

+++ src/main/java/com/topably/assets/portfolios/domain/cards/input/DividendGoalsCard.java
@@ -4,6 +4,7 @@ import com.topably.assets.portfolios.domain.cards.PortfolioCard;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
@@ -12,6 +13,6 @@ import java.util.Map;
 @Setter
 public class DividendGoalsCard extends PortfolioCard {
 
-    private Map<String, BigInteger> desiredYieldByIssuer = new HashMap<>();
+    private Map<String, BigDecimal> desiredYieldByIssuer = new HashMap<>();
 
 }
